requirements.txt
# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
pillow>=10.0.0
numpy>=1.24.0

# CLIP for verification
clip @ git+https://github.com/openai/CLIP.git

# Configuration management
pyyaml>=6.0

# Jupyter and analysis
jupyter>=1.0.0
ipython>=8.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
pandas>=1.5.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
